// src/app/Model/Pessoa.ts
export interface Pessoa {
    id: number;
    nome: string;
    cpf: string;
    telFixo: string;
    celular: string;
    dataNasc: Date;
    Email: string;
    Fb: string;
    insta: string;
    uf: string;
    cidade: string;
    complemento: string;
    bairro: string;
    rua: string;
    cep: string;
}